requirements.txt
# 图片管理器依赖包
Pillow>=9.0.0
psutil>=5.8.0
watchdog>=2.1.0
pyinstaller>=5.0.0
pywin32>=306; sys_platform == "win32"